# public/openApi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-messenger",
    "description": "Chat app",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/e-messenger",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/chat/histories/{conversationId}": {
      "get": {
        "tags": [
          "chatting-controller"
        ],
        "operationId": "getMessageHistory",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListMessageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseListMessageResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            }
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "senderName": {
            "type": "string"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "user token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}